Refactor CEF configuration into platform-specific files
